requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
requests==2.32.3
urllib3==2.2.3

allure-python-commons~=2.13.5
pytest~=8.3.4